Removed mongodb specific information from classes
