+++ .gitignore
@@ -133,3 +133,8 @@ dmypy.json
 
 # Testing
 .coverage
+
+# Secrets
+*.secret
+secret/
+secrets/
