+++ README.md
@@ -25,3 +25,7 @@ Complete list of Pokemon with stats (with forms - beware of darumaka it has both
 Complete list of Pokemon API with all details (optional variants or separate)
 
 Scrape based on time periods and handle all errors
+
+## Rationale for using Python 3.8
+
+Python 3.8 was used specifically to take advatange of typing.Final and some features from python 3.7 like dict order preservation.
