+++ requirements.txt
@@ -3,16 +3,30 @@ alabaster==0.7.12
 attrs==19.3.0
 Babel==2.7.0
 beautifulsoup4==4.8.1
+CacheControl==0.12.6
+cachetools==4.0.0
 certifi==2019.11.28
 cffi==1.13.2
 chardet==3.0.4
 coverage==4.5.4
 cryptography==2.8
 docutils==0.15.2
+firebase-admin==3.2.1
+google-api-core==1.15.0
+google-api-python-client==1.7.11
+google-auth==1.10.0
+google-auth-httplib2==0.0.3
+google-cloud-core==1.1.0
+google-cloud-firestore==1.6.0
+google-cloud-storage==1.23.0
+google-resumable-media==0.5.0
+googleapis-common-protos==1.6.0
+grpcio==1.26.0
 h11==0.8.1
 h2==3.1.1
 hpack==3.0.0
 httpcore==0.3.0
+httplib2==0.15.0
 httptools==0.0.13
 hyperframe==5.2.0
 idna==2.8
@@ -22,10 +36,14 @@ Jinja2==2.10.3
 loguru==0.4.0
 MarkupSafe==1.1.1
 more-itertools==8.0.2
+msgpack==0.6.2
 multidict==4.6.1
 packaging==19.2
 pluggy==0.13.1
+protobuf==3.11.2
 py==1.8.0
+pyasn1==0.4.8
+pyasn1-modules==0.2.7
 pycparser==2.19
 Pygments==2.5.2
 pymongo==3.10.0
@@ -37,6 +55,7 @@ pytz==2019.3
 requests==2.22.0
 requests-async==0.5.0
 rfc3986==1.3.2
+rsa==4.0
 sanic==19.9.0
 six==1.13.0
 snowballstemmer==2.0.0
@@ -49,6 +68,7 @@ sphinxcontrib-jsmath==1.0.1
 sphinxcontrib-qthelp==1.0.2
 sphinxcontrib-serializinghtml==1.1.3
 ujson==1.35
+uritemplate==3.0.1
 urllib3==1.25.7
 uvloop==0.12.2
 wcwidth==0.1.7
