+++ src/data/egg_group_list.py
@@ -1,6 +1,5 @@
 from dataclasses import dataclass, field, asdict
 from typing import List, Dict, Any
-from bson import ObjectId
 from src.data.poke_enums import EggGroup
 
 
@@ -8,8 +7,7 @@ from src.data.poke_enums import EggGroup
 class EggGroupList:
     """List of pokemon from a specific egg group"""
     group: EggGroup = EggGroup.INVALID
-    pokemon_list: List[str] = field(default_factory=lambda: ObjectId())
-    _id: ObjectId = field(default_factory=lambda: ObjectId())
+    pokemon_list: List[str] = field(default_factory=lambda: '')
 
     def asdict(self) -> Dict[str, Any]:
         return asdict(self)
