+++ src/data/pmove.py
@@ -1,5 +1,4 @@
 from dataclasses import dataclass, field, asdict
-from bson import ObjectId
 from typing import Dict, Any
 from src.data.poke_enums import PType, Target
 
@@ -9,10 +8,7 @@ class PMove:
     """Description of a move and its effects. """
     # TODO: Populate this
     name: str = ''
-
     base_power: int = 0
-
-    _id: ObjectId = field(default_factory=lambda: ObjectId())
 
     def asdict(self) -> Dict[str, Any]:
         return asdict(self)
@@ -69,8 +65,6 @@ class PMoveEssentials:
     # Flags end
 
     description: str = ''
-
-    _id: ObjectId = field(default_factory=lambda: ObjectId())
 
     def asdict(self) -> Dict[str, Any]:
         return asdict(self)
